requirements.txt
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0